+++ Frontend/sopsc-mobile-app/.env.example
@@ -1,0 +1,1 @@
+EXPO_PUBLIC_API_URL=https://2c87-137-119-8-154.ngrok-free.app/api/
