+++ Frontend/sopsc-mobile-app/src/config.ts
@@ -1,0 +1,1 @@
+export const API_URL = process.env.EXPO_PUBLIC_API_URL ?? '';
